requirements.txt
pytest
requests
sphinx
Mastodon.py